+++ services/horizon/internal/config.go
@@ -19,7 +19,6 @@ type Config struct {
 	Port               uint
 	AdminPort          uint
 
-	EnableCaptiveCoreIngestion  bool
 	EnableIngestionFiltering    bool
 	CaptiveCoreBinaryPath       string
 	RemoteCaptiveCoreURL        string
@@ -30,8 +29,7 @@ type Config struct {
 	CaptiveCoreReuseStoragePath bool
 	CaptiveCoreConfigUseDB      bool
 
-	StellarCoreDatabaseURL string
-	StellarCoreURL         string
+	StellarCoreURL string
 
 	// MaxDBConnections has a priority over all 4 values below.
 	MaxDBConnections            int